pyproject.toml
[tool.poetry]
name = "django-bookshelf"
version = "0.1.0"
description = ""
authors = ["Cleyson Lima <cleysonph@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
django = "^3.0.4"
python-decouple = "^3.3"
psycopg2-binary = "^2.8.4"
django-stdimage = "^5.0.2"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
